Visual - Fully convert enemy and player speeds to floats

+++ src/visual/entities/ventity_enemy.py
@@ -55,7 +55,7 @@ class VEntityEnemyCommon(VEntityMoving):
 
     # ########################################################################
     # ############################################################# SPEED ####
-    def get_speed(self) -> int:
+    def get_speed(self) -> float:
         return self.gamestate.enemy_speed
 
     def get_target_sprite(self) -> Sprite:
